+++ packages/core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- feat(live-server): add openPathOnStartup configuration for live server. (`039b74dc`)
 - refactor(teamHub): introduce soft-connection handling for team hubs. (`5fc744ec`)
 - feat(live-server): integrate live-server package and enhance CLI functionality. (`890d9cd7`)