+++ packages/sdk/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- feat(live-server): add openPathOnStartup configuration for live server. (`039b74dc`)
+
 ## 1.4.4 - 2026-08-01
 
 - refactor(teamHub): introduce soft-connection handling for team hubs. (`5fc744ec`)